Add debug log printing to messages
There will be redundancy when debug logging is turned on, but only if
the debug output is going to the console and not a file.  If the debug
output is being sent to a file, we need to print the messages.py
messages or else we never see them in the file.  The possible
duplication of messages is the price we pay for simplicity....