Add new version of CLI & GUI classes
This is another iteration of the message handler/CLI/GUI code I
developed for many of these applications.  This time I've split it out
into a separate subdirectory.

The GUI part is not used in Microarchiver but I'm leaving it in for
now, in case something changes in the future and we end up wanting to
make it operational.