Do not throw exception if fail to get data from DataCite
Looks like things were inconsistently written. In one place it
checked for missing DataCite data and simply flagged it, while the
part that interacted with DataCite threw an exception.  I made the
latter return None instead, so it should now keep going on the
remaining articles if it encounters a problem with getting metadata
from DataCite.