Fix last-run date handling
The value of the `--after-date` argument to `microarchiver` was set to the same date that it was run, which caused it to miss articles published on that date the next time it ran.  The value of the date should have been modified to include the day it last ran so that the "after date" date comparison was correct.  (Thanks to Nick Stiffler for catching and reporting the problem.)